Basic implementation of authentication and fetching some things
We keep a local file with our creds.
The first time we talk to an instance, we register our app and get our client creds.
The first time we try to log in as a particular user, we prompt for a password.

We attempt to be somewhat smart about defaults - if no user is specified but we have existing
creds for exactly one user, we use that user.